Move and rename `ActiveARViewScannerContainer` to `ActiveARViewContainer` file.
